+++ src/aux-server/aux-web/shared/vue-components/AuthUI/AuthUI.ts
@@ -50,6 +50,7 @@ export default class AuthUI extends Vue {
 
     showAccountInfo: boolean = false;
     loginStatus: LoginStatus = null;
+    supportUrl: string = null;
 
     get isLoggedIn() {
         return !!this.loginStatus;
@@ -82,6 +83,7 @@ export default class AuthUI extends Vue {
         this.showNotAuthorized = false;
         this.showAccountInfo = false;
         this.loginStatus = null;
+        this.supportUrl = null;
         this._sub = new Subscription();
 
         this._sub.add(
@@ -113,6 +115,15 @@ export default class AuthUI extends Vue {
                 this.showAccountInfo = true;
                 this.loginStatus = e.loginStatus;
                 this._simId = e.simulationId;
+
+                if (e.endpoint) {
+                    const endpoint = appManager.auth.getEndpoint(e.endpoint);
+                    endpoint.getPolicyUrls().then((urls) => {
+                        this.supportUrl = urls.supportUrl;
+                    });
+                } else {
+                    this.supportUrl = null;
+                }
 
                 this.reportInstVisible = false;
                 if (this._simId) {