+++ src/aux-server/aux-web/shared/vue-components/EnterAddressDialog/EnterAddressDialog.vue
@@ -48,6 +48,9 @@
                         <p v-if="termsOfServiceUrl">
                             <a target="_blank" :href="termsOfServiceUrl">Terms of Service</a>
                         </p>
+                        <p v-if="supportUrl">
+                            <a target="_blank" :href="supportUrl">Support</a>
+                        </p>
                     </div>
                 </div>
             </md-dialog-content>